requirements.txt
django>=4.0,<5.0
gunicorn>=20.0,<21.0
dj-database-url>=1.0,<2.0
whitenoise>=6.0,<7.0
psycopg2-binary>=2.0,<3.0
livereload>=2.6.3
numpy>=1.23.4
pandas>=1.5.0
openai>=0.25.0
resemble>=1.3.0
python-dotenv>=0.21.0
transformers>=4.25.1
pypdf2>=3.0.1
